[dvc][server] Added a DelegatingStorageEngine
For now, the only class extending this is the DeepCopyStorageEngine.

In a later commit, it will also be used in order to fix a bug in the
ingestion code.

This is a non-functional change.
